\name{tabix_close}
\alias{tabix_close}
\title{
Close Tabix-indexed file
}
\description{
Close Tabix-indexed file
}
\usage{
tabix_close(tabfh)
}
\arguments{
  \item{tabfh}{Tabix file handle}
}
\value{
None.
}
\author{
Ulrich Wittelsbuerger
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
tabix_open, tabix_read
}
