\name{whop.eg.abbrevForOrganism}
\alias{whop.eg.abbrevForOrganism}
\title{
Look up the organism prefix for the .org.eg.db databases from Bioconductor
}
\description{
Look up the organism prefix for the .org.eg.db databases from Bioconductor
}
\usage{
whop.eg.abbrevForOrganism(organismname)
}
\arguments{
  \item{organismname}{Name of organism}
}
\details{
Used internally.
}
\value{Database prefix}
\author{
Ulrich Wittelsbuerger
}
