\name{whop.eg.chromosome}
\alias{whop.eg.chromosome}
\title{
Return the chromosome on which the gene identified by the given Entrez ID lies.
}
\description{
Return the chromosome on which the gene identified by the given Entrez ID lies.
}
\usage{
whop.eg.chromosome(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Chromosome name
}
\author{
Ulrich Wittelsbuerger
}
