\name{whop.eg.eg_lookup}
\alias{whop.eg.eg_lookup}
\title{
Return all entries in an EG organism's data table for all given identifiers
}
\description{
Return all entries in an EG organism's data table for all given identifiers
}
\usage{
whop.eg.eg_lookup(ids, subdbname, db)
}
\arguments{
  \item{ids}{Identifiers to look for}
  \item{subdbname}{Subtable to look in}
  \item{db}{Organism's database if not using default currently active one}
}
\value{
Depends on table
}
\author{
Ulrich Wittelsbuerger
}
