\name{whop.eg.eg_lookupAll}
\alias{whop.eg.eg_lookupAll}
\title{
Return all entries in an EG organism's data table for a given identifier
}
\description{
Return all entries in an EG organism's data table for a given identifier
}
\usage{
whop.eg.eg_lookupAll(id, subdbname, db)
}
\arguments{
  \item{id}{Identifier(s) to look for in subtable}
  \item{subdbname}{Organism annotation table name}
  \item{db}{Optional, organism database if not using default active one}
}
\value{
Depends on table
}
\author{
Ulrich Wittelsbuerger
}
