\name{whop.eg.eg_RevLookup}
\alias{whop.eg.eg_RevLookup}
\title{
Perform a reverse lookup on one of the EG organism database's sub-tables.
}
\description{
Perform a reverse lookup on one of the EG organism database's sub-tables.
}
\usage{
whop.eg.eg_RevLookup(ids, subdbname, db)
}
\arguments{
  \item{ids}{Identifiers to look for in subtable}
  \item{subdbname}{Organism annotation table name}
  \item{db}{Optional, organism database if not using default active one}
}
\value{
Depends on data queried
}
\author{
Ulrich Wittelsbuerger
}
