\name{whop.eg.fromAccnum}
\alias{whop.eg.fromAccnum}
\title{
Turn a GenBank accession number into a Entrez identifier.
}
\description{
Turn a GenBank accession number into a Entrez identifier.
}
\usage{
whop.eg.fromAccnum(id, db)
}
\arguments{
  \item{id}{GenBank accession}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
