\name{whop.eg.fromAlias}
\alias{whop.eg.fromAlias}
\title{
Turn an Alias into a Entrez identifier.
}
\description{
Turn an Alias into a Entrez identifier.
}
\usage{
whop.eg.fromAlias(id, db)
}
\arguments{
  \item{id}{Alias}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
