\name{whop.eg.fromEnsembl}
\alias{whop.eg.fromEnsembl}
\title{
Turn an Ensembl identifier into a Entrez identifier.
}
\description{
Turn an Ensembl identifier into a Entrez identifier.
}
\usage{
whop.eg.fromEnsembl(id, db)
}
\arguments{
  \item{id}{Ensembl identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
