\name{whop.eg.fromEnsemblProt}
\alias{whop.eg.fromEnsemblProt}
\title{
Turn an Ensembl Protein identifier into a Entrez identifier.
}
\description{
Turn an Ensembl Protein identifier into a Entrez identifier.
}
\usage{
whop.eg.fromEnsemblProt(id, db)
}
\arguments{
  \item{id}{Ensembl Protein identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
