\name{whop.eg.fromEnsemblTrans}
\alias{whop.eg.fromEnsemblTrans}
\title{
Turn an Ensemble transcript identifier into a Entrez identifier.
}
\description{
Turn an Ensemble transcript identifier into a Entrez identifier.
}
\usage{
whop.eg.fromEnsemblTrans(id, db)
}
\arguments{
  \item{id}{Ensembl Transcript identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
