\name{whop.eg.fromEnzyme}
\alias{whop.eg.fromEnzyme}
\title{
Turn an Enzyme nomenclature identifier into a Entrez identifier.
}
\description{
Turn an Enzyme nomenclature identifier into a Entrez identifier.
}
\usage{
whop.eg.fromEnzyme(id, db)
}
\arguments{
  \item{id}{Enzyme EC identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
