\name{whop.eg.fromGO}
\alias{whop.eg.fromGO}
\title{
Turn a GO term identifier into a related Entrez identifier.
}
\description{
Turn a GO term identifier into a related Entrez identifier.
}
\usage{
whop.eg.fromGO(id, db)
}
\arguments{
  \item{id}{GO term identifier}
  \item{db}{Organism database to look in, if not using currently active one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
