\name{whop.eg.fromGO2AllEgs}
\alias{whop.eg.fromGO2AllEgs}
\title{
Return all Entrez identifiers related to a given GO term.
}
\description{
Return all Entrez identifiers related to a given GO term.
}
\usage{
whop.eg.fromGO2AllEgs(id, db)
}
\arguments{
  \item{id}{GO term identifier}
  \item{db}{Organism database to look in, if not using currently active one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
