\name{whop.eg.fromOmim}
\alias{whop.eg.fromOmim}
\title{
Turn an OMIM identifier into a Entrez identifier.
}
\description{
Turn an OMIM identifier into a Entrez identifier.
}
\usage{
whop.eg.fromOmim(id, db)
}
\arguments{
  \item{id}{OMIM identifier}
  \item{db}{Organism database to look in, if not using currently active one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
