\name{whop.eg.fromPath}
\alias{whop.eg.fromPath}
\title{
Turn a KEGG pathway identifier into related Entrez identifiers.
}
\description{
Turn a KEGG pathway identifier into related Entrez identifiers.
}
\usage{
whop.eg.fromPath(id, db)
}
\arguments{
  \item{id}{KEGG pathway identifier}
  \item{db}{Organism database to look in, if not using currently active one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
