\name{whop.eg.fromPmid}
\alias{whop.eg.fromPmid}
\title{
Turn an PMID identifier into a Entrez identifier.
}
\description{
Turn an PMID identifier into a Entrez identifier.
}
\usage{
whop.eg.fromPmid(id, db)
}
\arguments{
  \item{id}{PMID identifier}
  \item{db}{Organism database to look in, if not using currently active one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
