\name{whop.eg.fromRefseq}
\alias{whop.eg.fromRefseq}
\title{
Turn a Refseq identifier into a Entrez identifier.
}
\description{
Turn a Refseq identifier into a Entrez identifier.
}
\usage{
whop.eg.fromRefseq(id, db)
}
\arguments{
  \item{id}{Refseq identifier}
  \item{db}{Organism database to look in, if not using currently active one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
