\name{whop.eg.fromUnigene}
\alias{whop.eg.fromUnigene}
\title{
Turn an Unigene identifier into a Entrez identifier.
}
\description{
Turn an Unigene identifier into a Entrez identifier.
}
\usage{
whop.eg.fromUnigene(id, db)
}
\arguments{
  \item{id}{Unigene identifier}
  \item{db}{Organism database to look in, if not using currently active one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
