\name{whop.eg.fromUniprot}
\alias{whop.eg.fromUniprot}
\title{
Turn an Uniprot identifier into a Entrez identifier.
}
\description{
Turn an Uniprot identifier into a Entrez identifier.
}
\usage{
whop.eg.fromUniprot(id, db)
}
\arguments{
  \item{id}{Uniprot identifier}
  \item{db}{Organism database to look in, if not using currently active one}
}
\value{
Entrez identifier(s)
}
\author{
Ulrich Wittelsbuerger
}
