\name{whop.eg.genename}
\alias{whop.eg.genename}
\title{
Find the gene name for a given Entrez identifier
}
\description{
Find the gene name for a given Entrez identifier
}
\usage{
whop.eg.genename(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Gene names
}
\author{
Ulrich Wittelsbuerger
}
