\name{whop.eg.goIds}
\alias{whop.eg.goIds}
\title{
Returns GO term identifiers related to the given Entrez identifier.
}
\description{
Returns GO term identifiers related to the given Entrez identifier.
}
\usage{
whop.eg.goIds(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
GO identifiers
}
\author{
Ulrich Wittelsbuerger
}
