\name{whop.eg.installdb}
\alias{whop.eg.installdb}
\title{
Download and install the Bioconductor EG database for a given organism
}
\description{
Download and install the Bioconductor EG database for a given organism
}
\usage{
whop.eg.installdb(organismname)
}
\arguments{
  \item{organismname}{Organism name or abbreviation}
}
\details{
Attempts to automatically download and install an organism's annotation database from Bioconductor
}
\value{
Success status
}
\author{
Ulrich Wittelsbuerger
}
