\name{whop.eg.keggpathways}
\alias{whop.eg.keggpathways}
\title{
Look up KEGG pathway identifiers related to the given Entrez identifier.
}
\description{
Look up KEGG pathway identifiers related to the given Entrez identifier.
}
\usage{
whop.eg.keggpathways(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
KEGG PATHWAY identifiers
}
\author{
Ulrich Wittelsbuerger
}
