\name{whop.eg.load_orgdb}
\alias{whop.eg.load_orgdb}
\title{
Load and, if necessary, install a Bioconductor EG database for a given organism.
}
\description{
Load and, if necessary, install a Bioconductor EG database for a given organism.
}
\usage{
whop.eg.load_orgdb(organismname, install.if.missing = F)
}
\arguments{
  \item{organismname}{Organism name or abbreviation}
  \item{install.if.missing}{Install database if not present locally?}
}
\value{
Success status
}
\author{
Ulrich Wittelsbuerger
}
