\name{whop.eg.Organism}
\alias{whop.eg.Organism}
\title{
Returns the organism's name for which the current database-set contains information.
}
\description{
Returns the organism's name for which the current database-set contains information.
}
\usage{
whop.eg.Organism()
}
\value{
String: organism name
}
\author{
Ulrich Wittelsbuerger
}
