\name{whop.eg.orgdb_loaded}
\alias{whop.eg.orgdb_loaded}
\title{
Find out whether a certain organism's Bioconductor EG database has been loaded
}
\description{
Find out whether a certain organism's Bioconductor EG database has been loaded
}
\usage{
whop.eg.orgdb_loaded(organismname)
}
\arguments{
  \item{organismname}{Organism's name}
}
\value{
TRUE or FALSE
}
\author{
Ulrich Wittelsbuerger
}
