\name{whop.eg.region}
\alias{whop.eg.region}
\title{
Look up the start and end of the gene identified by the given Entrez ID.
}
\description{
Look up the start and end of the gene identified by the given Entrez ID.
}
\usage{
whop.eg.region(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Start and end positions
}
\author{
Ulrich Wittelsbuerger
}
