\name{whop.eg.toAccnum}
\alias{whop.eg.toAccnum}
\title{
Look up for an Entrez identifier the corresponding GenBank Accession number.
}
\description{
Look up for an Entrez identifier the corresponding GenBank Accession number.
}
\usage{
whop.eg.toAccnum(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Translated identifiers
}
\author{
Ulrich Wittelsbuerger
}
