\name{whop.eg.toAlias}
\alias{whop.eg.toAlias}
\title{
Look up the corresponding common alias for an Entrez identifier.
}
\description{
Look up the corresponding common alias for an Entrez identifier.
}
\usage{
whop.eg.toAlias(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Translated identifiers
}
\author{
Ulrich Wittelsbuerger
}
