\name{whop.eg.toEnsemblProt}
\alias{whop.eg.toEnsemblProt}
\title{
Look up for an Entrez identifier the corresponding Ensembl Protein identifiers.
}
\description{
Look up for an Entrez identifier the corresponding Ensembl Protein identifiers.
}
\usage{
whop.eg.toEnsemblProt(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Translated identifiers
}
\author{
Ulrich Wittelsbuerger
}
