\name{whop.eg.toGO}
\alias{whop.eg.toGO}
\title{
Look up for an Entrez identifier the corresponding GO terms.
}
\description{
Look up for an Entrez identifier the corresponding GO terms.
}
\usage{
whop.eg.toGO(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Translated identifiers
}
\author{
Ulrich Wittelsbuerger
}
