\name{whop.eg.toOmim}
\alias{whop.eg.toOmim}
\title{
Look up the OMIM identifier(s) corresponding to an Entrez identifier
}
\description{
Look up the OMIM identifier(s) corresponding to an Entrez identifier
}
\usage{
whop.eg.toOmim(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Translated identifiers
}
\author{
Ulrich Wittelsbuerger
}
