\name{whop.eg.toPmid}
\alias{whop.eg.toPmid}
\title{
Look up the Uniprot identifier(s) corresponding to an Entrez identifier
}
\description{
Look up the Uniprot identifier(s) corresponding to an Entrez identifier
}
\usage{
whop.eg.toPmid(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Translated identifiers
}
\author{
Ulrich Wittelsbuerger
}
