\name{whop.eg.toRefseq}
\alias{whop.eg.toRefseq}
\title{
Look up the Refseq identifier(s) corresponding to an Entrez identifier
}
\description{
Look up the Refseq identifier(s) corresponding to an Entrez identifier
}
\usage{
whop.eg.toRefseq(id, db)
}
\arguments{
  \item{id}{Entrez identifier}
  \item{db}{Organism database name, if not using currently activated one}
}
\value{
Translated identifiers
}
\author{
Ulrich Wittelsbuerger
}
