\name{whop.go.all_genes_for_term}
\alias{whop.go.all_genes_for_term}
\title{
Returns all genes related to the given GO term
}
\description{
Returns all genes related to the given GO term
}
\usage{
whop.go.all_genes_for_term(tomatch)
}
\arguments{
  \item{tomatch}{GO term name}
}
\value{
Genes
}
\author{
Ulrich Wittelsbuerger
}
