\name{whop.go.goid_like}
\alias{whop.go.goid_like}
\title{
Return GO terms with identifiers typographically similar to the given one
}
\description{
Return GO terms with identifiers typographically similar to the given one
}
\usage{
whop.go.goid_like(idmatch)
}
\arguments{
  \item{idmatch}{GO term}
}
\value{
GO terms
}
\author{
Ulrich Wittelsbuerger
}
