\name{whop.go.is_obsolete_byid}
\alias{whop.go.is_obsolete_byid}
\title{
Check obsolescence of GO terms with similar accessions
}
\description{
Returns all obsolete GO terms with similar accession
}
\usage{
whop.go.is_obsolete_byid(idmatch)
}
\arguments{
  \item{idmatch}{accession}
}
\value{
GO terms
}
\author{
Ulrich Wittelsbuerger
}
