\name{whop.go.is_obsolete_byname}
\alias{whop.go.is_obsolete_byname}
\title{
Check obsolescence of GO terms with similar names
}
\description{
Check obsolescence of GO terms with similar names
}
\usage{
whop.go.is_obsolete_byname(tomatch)
}
\arguments{
  \item{tomatch}{GO term name}
}
\value{
All obsolete GO terms matching the description
}
\author{
Ulrich Wittelsbuerger
}
