\name{whop.go.load}
\alias{whop.go.load}
\title{
Load a GO term database from file
}
\description{
Load a GO term database from file
}
\usage{
whop.go.load(filename = NA)
}
\arguments{
  \item{filename}{Filename of a GO database}
}
\value{
TRUE if any data has been read, FALSE if not
}
\author{
Ulrich Wittelsbuerger
}
