\name{whop.go.term_ancestors}
\alias{whop.go.term_ancestors}
\title{
Returns all ancestors of the given GO term.
}
\description{
Returns all ancestors of the given GO term.
}
\usage{
whop.go.term_ancestors(tomatch)
}
\arguments{
  \item{tomatch}{GO term}
}
\value{
GO terms
}
\author{
Ulrich Wittelsbuerger
}
