\name{whop.go.term_ancestors_similar}
\alias{whop.go.term_ancestors_similar}
\title{
Return ancestral GO terms of similarly named GO term.
}
\description{
For all GO terms named like the given, returns ancestral GO terms
}
\usage{
whop.go.term_ancestors_similar(tomatch)
}
\arguments{
  \item{tomatch}{GO term}
}
\value{
GO terms
}
\author{
Ulrich Wittelsbuerger
}
