\name{whop.go.term_synonyms}
\alias{whop.go.term_synonyms}
\title{
Returns GO terms synonymous with the given term
}
\description{
Returns GO terms synonymous with the given term
}
\usage{
whop.go.term_synonyms(tomatch)
}
\arguments{
  \item{tomatch}{GO term}
}
\value{
GO terms
}
\author{
Ulrich Wittelsbuerger
}
