\name{whop.go.terms_match}
\alias{whop.go.terms_match}
\title{
Returns all terms with names similar to the given one.
}
\description{
Returns the results of a SQL statement that extracts all terms with similar name to the given one.
}
\usage{
whop.go.terms_match(tomatch)
}
\arguments{
  \item{tomatch}{term}
}
\value{
GO terms
}
\author{
Ulrich Wittelsbuerger
}
