\name{whop.kegg.pathway_url}
\alias{whop.kegg.pathway_url}
\title{
Produces a URL to the KEGG website for a certain pathway
}
\description{
For all KEGG pathway IDs given, a URL to the KEGG webpage for that pathway is returned.
}
\usage{
whop.kegg.pathway_url(pathwayids)
}
\arguments{
  \item{pathwayids}{One or more KEGG pathway identifiers}
}
\value{
A string containing an URL or vector of URLs
}
\author{
Ulrich Wittelsbuerger
}
