\name{whop.ped.entriesOf}
\alias{whop.ped.entriesOf}
\title{
Return all entries from a pedigree dataset matching the list of given identifiers.
}
\description{
Returns pedigree data on all individuals given in parameter 2
}
\usage{
whop.ped.entriesOf(p, invids)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p}{The pedigree dataset}
  \item{invids}{The identifiers of the individuals to extract}
}
\value{
Table of rows from the pedigree
}
\author{
Ulrich Wittelsbuerger
}
