\name{whop.ped.familyOf}
\alias{whop.ped.familyOf}
\title{
Returns all members of an individuals family
}
\description{
Returns all members of an individuals family
}
\usage{
whop.ped.familyOf(p, lis)
}
\arguments{
  \item{p}{The pedigree dataset}
  \item{lis}{The individual(s) for which family members should be extracted}
}
\value{
Table of rows from the pedigree
}
\author{
Ulrich Wittelsbuerger
}
