\name{whop.ped.fathers}
\alias{whop.ped.fathers}
\title{
Return all fathers from a pedigree dataset
}
\description{
Returns pedigree data on all individuals which appear in the Paternal.ID column
}
\usage{
whop.ped.fathers(p)
}
\arguments{
  \item{p}{The pedigree dataset}
}
\value{
Table of rows from the pedigree
}
\author{
Ulrich Wittelsbuerger
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
whop.ped.mothers
}
