\name{whop.ped.females}
\alias{whop.ped.females}
\title{
Return all females from a pedigree dataset
}
\description{
Extracts all individuals with 'Sex' defined as female
}
\usage{
whop.ped.females(p)
}
\arguments{
  \item{p}{The pedigree dataset}
}
\value{
Table of rows from the pedigree
}
\author{
Ulrich Wittelsbuerger
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
whop.ped.males
}
