\name{whop.ped.fromPop}
\alias{whop.ped.fromPop}
\title{
Return all individuals belonging to a given population
}
\description{
All individuals with one of the given population IDs are returned as a pedigree table.
}
\usage{
whop.ped.fromPop(p, popids)
}
\arguments{
  \item{p}{The pedigree dataset}
  \item{popids}{A vector with one or more population IDs}
}
\value{
Table of rows from the pedigree
}
\author{
Ulrich Wittelsbuerger
}
