\name{whop.ped.males}
\alias{whop.ped.males}
\title{
Return only the male individuals from a pedigree dataset
}
\description{
Extract all male individuals from a pedigree dataset that has been previously loaded with whop.ped.load()
}
\usage{
whop.ped.males(p)
}
\arguments{
  \item{p}{The pedigree dataset}
}
\value{
Table of rows from the pedigree
}
\author{
Ulrich Wittelsbuerger
}
\seealso{
\code{\link{whop.ped.females}}
\code{\link{bgzf_compress}}
}
