\name{whop.ped.mothers}
\alias{whop.ped.mothers}
\title{
Get all mothers stored in a pedigree file
}
\description{
All individuals which appear in the Maternal.ID column of the pedigree data
}
\usage{
whop.ped.mothers(p)
}
\arguments{
  \item{p}{The pedigree dataset}
}
\value{
Table of rows from the pedigree
}
\author{
Ulrich Wittelsbuerger
}
\seealso{
whop.ped.fathers
}
