\name{whop.ped.names}
\alias{whop.ped.names}
\title{
Get all individual names
}
\description{
Returns a vector of strings, containing all Individual.IDs from the pedigree data
}
\usage{
whop.ped.names(p)
}
\arguments{
  \item{p}{The pedigree dataset}
}
\value{
A vector of strings, containing all Individual.IDs from the pedigree data
}
\author{
Ulrich Wittelsbuerger
}
