\name{whop.ped.save}
\alias{whop.ped.save}
\title{
Save pedigree data to file
}
\description{
Saves the pedigree dataset in p to a file.
}
\usage{
whop.ped.save(p, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p}{The pedigree dataset}
  \item{filename}{Name of the file to save into}
}
\value{
None.
}
\author{
Ulrich Wittelsbuerger
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
whop.ped.load()
}
