\name{whop.ucsc.query}
\alias{whop.ucsc.query}
\title{
Send a SQL query string to the UCSC Genome Browser SQL server
}
\description{
	The items given as parameters are concatenated into a SQL query string and sent to the
	UCSC Genome Browser SQL server.
}
\usage{
	whop.ucsc.query(...)
}
\arguments{
  \item{\dots}{any number of strings and variables that will be pasted together to build the query string}
}
\value{
	The returned value(s) from the UCSC Genome Browser.
}
%XXXXXXXXXXXXXXX
\examples{

##
##	Example :
##

}
%XXXXXXXXXXXXXXX
\author{
Ulrich Wittelsbuerger
}
